+++ src/rss/utils/fake.py
@@ -4,8 +4,8 @@ import random
 from sqlalchemy import create_engine, text
 from sqlalchemy.orm import sessionmaker
 
-from rss.api.config import APIConfig
-from rss.api import create_app, db
+from rss.api.config import DevAPIConfig, ProdAPIConfig
+from rss.api import create_app, db, API_CONFIG
 from rss.api.models import Alert, State
 from rss.services import MessageProcessor
 from rss.cap.states import STATES
@@ -13,7 +13,7 @@ from rss.cap.regions import REGIONS
 
 
 def clear_database() -> None:
-    engine = create_engine(APIConfig.SQLALCHEMY_DATABASE_URI)
+    engine = create_engine(DevAPIConfig.SQLALCHEMY_DATABASE_URI)
     engine.connect()
     session = sessionmaker(bind=engine)()
 
@@ -26,9 +26,14 @@ def clear_database() -> None:
 def generate_fake_alerts():
     """ Generate fake alerts and put then in the database.
     """
+    if "prod" in API_CONFIG.NAME or isinstance(API_CONFIG, ProdAPIConfig):
+        print("Cannot generate alerts in production mode. Exiting...")
+        return
+
+    print("Adding new alerts to DB")
     clear_database()
 
-    app = create_app(APIConfig())
+    app = create_app(DevAPIConfig())
 
     end_date = datetime.date.today()
     end_date = datetime.datetime(year=end_date.year, month=end_date.month, day=end_date.day)
@@ -103,9 +108,9 @@ def generate_fake_alerts():
         print(f"Total number of alerts {n_alerts}")
 
     print(id_list)
+    print("DONE")
 
 
 if __name__ == "__main__":
-    print("Adding new alerts to DB")
     generate_fake_alerts()
-    print("DONE")
+